Refactor everything, implement multi-threading, implement step-wise training approach (approach a value instead of starting at that value)
